Move `get_inf_range` from _util to drop_inf
